llm_gateway/requirements.txt
litellm[proxy]
uvicorn
python-dotenv
fastapi 